+++ src/main-extension-api.js
@@ -2,6 +2,7 @@ import { log, mod } from './utils.js';
 import { get_print_index_by_date } from './get_print_index_by_date.js';
 import { apply_to_buttons } from './apply_to_buttons.js';
 import { apply_to_elements } from './apply_to_elements.js';
+import { parse_color } from './parse_color.js';
 
 const COLOR = {
     autoset: true,