+++ changelog.md
@@ -1,7 +1,36 @@
 # pending changes
 
 
-* Can set an array of COLOR Objects to allow for more than one update state of the system
+* can define an array of COLOR objects to use with the extension
+
+```
+const COLOR_DEFAULT = {
+    autoset: true,
+    debug: false,
+    first_tuesday: new Date(2025, 9 - 1, 9, 0, 0, 0, 0),
+    first_index: 0,
+    ascending: true,
+    data: [  
+        { i: 0, desc: 'Green',  web: '#00ff00' },
+        { i: 1, desc: 'Blue',   web: '#0000ff' },
+        { i: 2, desc: 'Yellow', web: '#ffff00' },
+        { i: 3, desc: 'Orange', web: '#ff8800' },
+        { i: 4, desc: 'Red',    web: '#ff0000' }
+    ],
+    color: 'Green'
+};
+const COLOR_ARRAY_DEFAULT = [ COLOR_DEFAULT ];
+
+const parse_color_object = (COLOR) => {
+    return Object.assign({}, COLOR_DEFAULT, COLOR);
+};
+
+const get_color_array = () => {
+    return COLOR_ARRAY_DEFAULT;
+};
+
+```
+
 * RCTF.createColor method that will return a new COLOR Object
 * RCTF.setLocalConfig api method to set the local level config